requirements.txt
discord.py
python-dotenv
fuzzywuzzy
python-Levenshtein